+++ lib/auth-client.ts
@@ -2,13 +2,18 @@
 
 import { createAuthClient } from "better-auth/react";
 import {
+  genericOAuthClient,
   inferAdditionalFields,
   usernameClient,
 } from "better-auth/client/plugins";
 import type { auth } from "@/lib/auth";
 
 export const authClient = createAuthClient({
-  plugins: [inferAdditionalFields<typeof auth>(), usernameClient()],
+  plugins: [
+    inferAdditionalFields<typeof auth>(),
+    usernameClient(),
+    genericOAuthClient(),
+  ],
 });
 
 export const { useSession, signIn, signOut, getSession } = authClient;